+++ password_policies/admin.py
@@ -1,4 +1,5 @@
 from django.contrib import admin
+
 try:
     from django.utils.translation import gettext_lazy as _
 except ImportError:
@@ -18,6 +19,7 @@ def force_password_change(modeladmin, request, queryset):
     )
 
 
+@admin.register(PasswordHistory)
 class PasswordHistoryAdmin(admin.ModelAdmin):
     date_hierarchy = "created"
     exclude = ("password",)
@@ -33,6 +35,7 @@ class PasswordHistoryAdmin(admin.ModelAdmin):
         return False
 
 
+@admin.register(PasswordChangeRequired)
 class PasswordChangeRequiredAdmin(admin.ModelAdmin):
     date_hierarchy = "created"
     list_display = ("id", "user", "created")
@@ -51,7 +54,3 @@ class PasswordChangeRequiredAdmin(admin.ModelAdmin):
             return ["user"]
         else:
             return []
-
-
-admin.site.register(PasswordHistory, PasswordHistoryAdmin)
-admin.site.register(PasswordChangeRequired, PasswordChangeRequiredAdmin)