+++ HISTORY.rst
@@ -1,3 +1,16 @@
+Unreleased
+----------
+
+* Remove unsupported Django code
+* Added ``__str__()`` definitions for models
+* Use path instead of re_path for some URLs
+
+0.8.6
+-----
+
+* Django 5 support
+* default to JSONSerializer
+
 0.8.4
 -----